started ratings system

+++ src/screens/clientScreens/ServicesScreen/ServiceDetailsScreen.js
@@ -13,14 +13,18 @@ import {
 import styles from "./../../../../components/styles";
 import { firebase } from "../../../firebase/config";
 import useUser from "../../../../useUser";
+import Ratings from "../../../../components/Ratings";
 
 export default function ServiceDetailsScreen({ route, navigation }) {
   const [providerData, setProviderData] = useState(Object);
+  const [reviews, setReviews] = useState(Array);
   const { item } = route.params;
 
-  const { user } = useUser(); 
-  var docRef = firebase.firestore().collection("users/"+user.id+"/ClientFavorites");
-  
+  const { user } = useUser();
+  var docRef = firebase
+    .firestore()
+    .collection("users/" + user.id + "/ClientFavorites");
+
   useEffect(() => {
     firebase
       .firestore()
@@ -30,47 +34,66 @@ export default function ServiceDetailsScreen({ route, navigation }) {
       .then((querySnapshot) => {
         setProviderData(querySnapshot.data());
       });
+    getReviews();
   }, []);
 
+  const getReviews = () => {
+    // firebase.firestore
+    //   .collection("reviews")
+    //   .doc(item.providerId)
+    //   .get()
+    //   .then((querySnapshot) => {});
+    setReviews([
+      { rating: 3.5, review: "This is a review." },
+      { rating: 3.5, review: "This is a review." },
+    ]);
+  };
+
   const contactTel = () => {
-    if(checkAvailable(item.fromTime, item.toTime, item.contact) != ""){
+    if (checkAvailable(item.fromTime, item.toTime, item.contact) != "") {
       Linking.openURL(`tel:${item.contact}`);
     } else {
       return;
     }
-  }
+  };
   const onAddPress = () => {
-    docRef.doc(item.id).get().then((docSnapshot) => {
-      if (!docSnapshot.exists) {
-            firebase
-              .firestore()
-              .collection("users/"+user.id+"/ClientFavorites")
-              .doc(item.id)
-              .set(item)
-              .then(() => {
-                Alert.alert('Service has been favorited!')
-                navigation.navigate("Client Home");
-              });
-      } else {
-        Alert.alert('Service is already favorited.')
-      }
-    });
-  }
+    docRef
+      .doc(item.id)
+      .get()
+      .then((docSnapshot) => {
+        if (!docSnapshot.exists) {
+          firebase
+            .firestore()
+            .collection("users/" + user.id + "/ClientFavorites")
+            .doc(item.id)
+            .set(item)
+            .then(() => {
+              Alert.alert("Service has been favorited!");
+              navigation.navigate("Client Home");
+            });
+        } else {
+          Alert.alert("Service is already favorited.");
+        }
+      });
+  };
   const onUnfavoritePress = () => {
-    docRef.doc(item.id).get().then((docSnapshot) => {
-      if (docSnapshot.exists) {
-        docRef.doc(item.id).delete();
-        /* setlistData( listData => {
+    docRef
+      .doc(item.id)
+      .get()
+      .then((docSnapshot) => {
+        if (docSnapshot.exists) {
+          docRef.doc(item.id).delete();
+          /* setlistData( listData => {
           return listData.filter(Aservice => Aservice.id != item.id);
         }); */
-        Alert.alert('Service has been unfavorited.')
-        navigation.navigate("Client Home");
-      } else {
-        Alert.alert('Service was not previously favorited.')
-      }
-    }); 
-  }
-
+          Alert.alert("Service has been unfavorited.");
+          navigation.navigate("Client Home");
+        } else {
+          Alert.alert("Service was not previously favorited.");
+        }
+      });
+  };
+  console.log(reviews);
   return (
     <ImageBackground
       source={require("../../../../images/grey_background.png")}
@@ -81,22 +104,28 @@ export default function ServiceDetailsScreen({ route, navigation }) {
         <View style={styles.container}>
           <Text style={styles.title}>{item.serviceType}</Text>
           <Text style={styles.description}>{item.description}</Text>
-          <Text style={{ fontSize: 15, color: "#FEFEFE", marginBottom: 20}}>
-            Telephone Availability: {`${convertTo12Hour(item.fromTime)}`} - {`${convertTo12Hour(item.toTime)}`}  
+          <Text style={{ fontSize: 15, color: "#FEFEFE", marginBottom: 20 }}>
+            Telephone Availability: {`${convertTo12Hour(item.fromTime)}`} -{" "}
+            {`${convertTo12Hour(item.toTime)}`}
           </Text>
         </View>
         <View style={styles.locationNumberContainer}>
           <Text style={styles.location}>{item.location}</Text>
-          <Text style={styles.email}>{item.email}</Text> 
+          <Text style={styles.email}>{item.email}</Text>
         </View>
         <View>
-          <Text
-            style={styles.phoneNumber}
-            onPress={() => contactTel()}
-          >
+          <Text style={styles.phoneNumber} onPress={() => contactTel()}>
             {`${checkAvailable(item.fromTime, item.toTime, item.contact)}`}
           </Text>
-          <Text style={{fontSize: 15, color: "#FEFEFE", marginStart: 20, marginBottom: 20, justifyContent: 'center'}}>
+          <Text
+            style={{
+              fontSize: 15,
+              color: "#FEFEFE",
+              marginStart: 20,
+              marginBottom: 20,
+              justifyContent: "center",
+            }}
+          >
             {`${promptOutOfHours(item.fromTime, item.toTime, item.contact)}`}
           </Text>
         </View>
@@ -124,23 +153,61 @@ export default function ServiceDetailsScreen({ route, navigation }) {
             </View>
           </View>
           <View style={styles.profileDescriptionWrapper}>
-            <View>
+            <View
+              style={{ flexDirection: "row", justifyContent: "space-between" }}
+            >
               <Text style={styles.titleText}>Reviews</Text>
+              <Ratings reviews={reviews} />
             </View>
             <View style={styles.marginTop10}>
-              <Text>This Service Provider has no reviews yet.</Text>
+              {reviews.length == 0 && (
+                <Text>This Service Provider has no reviews yet.</Text>
+              )}
+              {reviews.length > 0 &&
+                reviews.map((item) => {
+                  return (
+                    <View
+                      style={{
+                        borderWidth: 1,
+                        borderStyle: "solid",
+                        borderColor: "#000",
+                        padding: 10,
+                        borderRadius: 5,
+                        marginBottom: 10,
+                      }}
+                    >
+                      <Text>{item.review}</Text>
+                      <Text>{item.rating}/5.0</Text>
+                    </View>
+                  );
+                })}
             </View>
           </View>
           <View style={styles.profileDescriptionWrapper}>
-            <View style={{flexDirection: 'row', justifyContent: 'center'}}>
+            <View style={{ flexDirection: "row", justifyContent: "center" }}>
               <TouchableOpacity
-                      style={{backgroundColor: "#FFAC1C", marginEnd: 20, width: 90, height: 30, alignItems: 'center',justifyContent: 'center'}}
-                      onPress={onAddPress}>
+                style={{
+                  backgroundColor: "#FFAC1C",
+                  marginEnd: 20,
+                  width: 90,
+                  height: 30,
+                  alignItems: "center",
+                  justifyContent: "center",
+                }}
+                onPress={onAddPress}
+              >
                 <Text style={styles.buttonTitle}>Favorite</Text>
               </TouchableOpacity>
               <TouchableOpacity
-                      style={{backgroundColor: "#FFAC1C", width: 90, height: 30, alignItems: 'center', justifyContent: 'center'}}
-                      onPress={onUnfavoritePress}>
+                style={{
+                  backgroundColor: "#FFAC1C",
+                  width: 90,
+                  height: 30,
+                  alignItems: "center",
+                  justifyContent: "center",
+                }}
+                onPress={onUnfavoritePress}
+              >
                 <Text style={styles.buttonTitle}>Unfavorite</Text>
               </TouchableOpacity>
             </View>
@@ -151,34 +218,34 @@ export default function ServiceDetailsScreen({ route, navigation }) {
   );
 }
 
-function promptOutOfHours(fromTime, toTime, contact){
-  if(checkAvailable(fromTime, toTime, contact) == ""){
-    return "Check back later during the allotted availability times for the provider's phone number!"
+function promptOutOfHours(fromTime, toTime, contact) {
+  if (checkAvailable(fromTime, toTime, contact) == "") {
+    return "Check back later during the allotted availability times for the provider's phone number!";
   } else {
     return "";
   }
 }
 
-function convertTo12Hour (time){
-  if (time < 13 && time > 0){
-    return (time).toString() + " A.M.";
+function convertTo12Hour(time) {
+  if (time < 13 && time > 0) {
+    return time.toString() + " A.M.";
   } else if (time > 12) {
-    return (time-12).toString() + " P.M."
+    return (time - 12).toString() + " P.M.";
   } else {
-    return "1 A.M."
+    return "1 A.M.";
   }
 }
 
-function checkAvailable (fromTime, toTime, tel) {
-  let currentHour = new Date().getHours();;
+function checkAvailable(fromTime, toTime, tel) {
+  let currentHour = new Date().getHours();
 
-  if(toTime < fromTime){
-    toTime = toTime+24;
+  if (toTime < fromTime) {
+    toTime = toTime + 24;
   }
-  if(currentHour < fromTime) {
-    currentHour = currentHour+24;
+  if (currentHour < fromTime) {
+    currentHour = currentHour + 24;
   }
-  if (toTime == fromTime){
+  if (toTime == fromTime) {
     return tel;
   } else if (currentHour > fromTime && currentHour < toTime) {
     return tel;
